Rename lib/index.js to lib/git-log.js
