+++ lib/git-log.js
